Added the controllers
